Added base_url, title, and redirect to database. Added useful_functions.
Onions.base_url is the root of the domain. Onions.title is the title of the base_url. Urls.redirect is the url to which a url redirects, if the url redirects at all.

The useful_functions.py file moves some functions from parse_scans.py into a more useful place, so that other files can access them.